+++ packages/external-ui-react/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename `Input` to `TextField`. `Input` still works.
+- `TextField` error state icon from an X to an alert triangle.
 
 ## [1.6.0] - 2019-02-22